+++ src/utils/hmac-validator.ts
@@ -10,10 +10,13 @@ import safeCompare from './safe-compare';
 export function stringifyQuery(query: AuthQuery): string {
   const orderedObj = Object.keys(query)
     .sort((val1, val2) => val1.localeCompare(val2))
-    .reduce((obj: Record<string, string | undefined>, key: keyof AuthQuery) => {
-      obj[key] = query[key];
-      return obj;
-    }, {});
+    .reduce(
+      (obj: {[key: string]: string | undefined}, key: keyof AuthQuery) => ({
+        ...obj,
+        [key]: query[key],
+      }),
+      {},
+    );
   return querystring.stringify(orderedObj);
 }
 
@@ -29,7 +32,7 @@ export function generateLocalHmac({
     timestamp,
     state,
     shop,
-    ...host && {host},
+    ...(host && {host}),
   });
   return crypto
     .createHmac('sha256', Context.API_SECRET_KEY)